requirements.txt
# meeting-hub Phase 1 依存

# ─── 音声取り込み／前処理 ───
faster-whisper>=1.1.0
whisperx>=3.1.0
librosa>=0.10.0
soundfile>=0.12.0
noisereduce>=3.0.0

# ─── 話者分離（pyannote provider） ───
pyannote.audio>=3.1.0

# ─── LLM（llm_cleanup / minutes_extract provider） ───
anthropic>=0.40.0

# ─── PyTorch（CPU 既定、GPU は別手順） ───
torch>=2.0.0
torchaudio>=2.0.0

# ─── スキーマ／設定／テンプレ ───
pydantic>=2.5
PyYAML>=6.0
jinja2>=3.1
python-dotenv>=1.0.0

# ─── Google Drive（サービスアカウント、§12-5） ───
google-api-python-client>=2.120.0
google-auth>=2.28.0

# ─── ライブ音声キャプチャ（Phase 2） ───
sounddevice>=0.4.6
scipy>=1.11

# ─── 外部連携（Phase 2） ───
notion-client>=2.2
slack-sdk>=3.27

# ─── Web UI / Runtime（Phase 3） ───
streamlit>=1.35
modal>=0.64

# ─── UX／ユーティリティ ───
tqdm>=4.66.0
